backend/requirements.txt
alembic==1.16.4
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
# psycopg2==2.9.10